Post Wilsonville competition commit.
Added many more autonomous modes; the robot can now place gears on the
left, center, and middle pegs. In addition, there is an option to make
the robot strafe to the boiler on either side after dropping off a gear
in autonomous. In the past, there have been issues regarding the gear
door sometimes passing over the limit switch without triggering it or
without stopping in time. The code now prevents this from happening by
looking at the last direction the door moved to figure out if it's open
or closed instead of waiting for a limit switch to be hit. The Rotate
command now has an option to rotate to an absolute angle instead of an
angle relative to the robot's current rotation. This is useful for a
consistent autonomous. The shooter subsystem now waits for the shooter
wheel to reach top speed before starting the spinner and staging wheel.
Improved overall tape detection in the vision system. Added a timeout
parameter for the gear alignment command, which allows you to specify
the max amount of time that should be spent aligning with the tape
before the command ends. This is useful for autonomous.